requirements.txt
bcrypt==4.0.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.2
decopatch==1.4.10
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
ecdsa==0.18.0
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.2
lxml==4.9.3
makefun==1.15.1
MarkupSafe==2.1.3
packaging==23.1
paramiko==3.3.1
pyasn1==0.5.0
pycparser==2.21
PyNaCl==1.5.0
python-gvm==24.8.0
python-jose==3.3.0
PyYAML==6.0.1
requests==2.31.0
requests-toolbelt==1.0.0
resolvelib==0.8.1
rsa==4.9
six==1.16.0
typing-extensions==4.7.1
urllib3==2.0.4
valid8==5.1.2
validators==0.20.0
xmltodict==0.13.0
zipp==3.16.2
netaddr==0.8.0
icalendar==5.0.7